requirements.txt
flask
mysql-connector-python
openai-whisper
pyannote.audio
python-dotenv
flask-mail
xhtml2pdf
sqlalchemy
pydub
torch
torchaudio
huggingface-hub
rouge-score